Change type of id, in Interface and stub class UserRepository, from int to Long

+++ src/main/java/com/busir/gardarian/bankloansystem/service/interfaces/UserRepositoryImpl.java
@@ -3,7 +3,7 @@ package com.busir.gardarian.bankloansystem.service.interfaces;
 import com.busir.gardarian.bankloansystem.entity.User;
 
 public interface UserRepositoryImpl {
-    public User findById(int id);
+    public User findById(Long id);
     public User findByUsername(String username);
     public User findByEmail(String email);
     public User save(User user);
